Remove all file I/O from Batfish.java
- move utility functions only used by internal CLI tools to another
class
- clean up CLI tools so they no longer reference Batfish
  - move some static functions as necessary
