+++ projects/batfish/src/test/java/org/batfish/main/FlattenTest.java
@@ -1,11 +1,16 @@
 package org.batfish.main;
 
+import static com.google.common.io.MoreFiles.createParentDirectories;
+import static org.batfish.common.util.CommonUtil.readFile;
+import static org.batfish.common.util.CommonUtil.readResource;
+import static org.batfish.common.util.CommonUtil.writeFile;
+import static org.batfish.main.Flatten.main;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
 
+import java.io.IOException;
 import java.nio.file.Path;
 import org.batfish.common.BfConsts;
-import org.batfish.common.util.CommonUtil;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -17,42 +22,41 @@ public final class FlattenTest {
   @Rule public ExpectedException _thrown = ExpectedException.none();
 
   @Test
-  public void testMainBadArgs() {
+  public void testMainBadArgs() throws IOException {
     _thrown.expect(IllegalArgumentException.class);
-    Flatten.main(new String[] {});
+    main(new String[] {});
   }
 
   private static final String TESTCONFIGS_PATH = "org/batfish/main/testconfigs";
 
-  public void assertInputOutputPair(String inputFilename, String outputFilename) {
+  public void assertInputOutputPair(String inputFilename, String outputFilename)
+      throws IOException {
     Path root = _folder.getRoot().toPath();
     Path inputDir = root.resolve("input");
     Path outputDir = root.resolve("output");
     Path inputFile = inputDir.resolve(BfConsts.RELPATH_CONFIGURATIONS_DIR).resolve("conf");
     Path outputFile = outputDir.resolve(BfConsts.RELPATH_CONFIGURATIONS_DIR).resolve("conf");
-    inputFile.getParent().toFile().mkdirs();
-    CommonUtil.writeFile(
-        inputFile,
-        CommonUtil.readResource(String.format("%s/%s", TESTCONFIGS_PATH, inputFilename)));
-    Flatten.main(new String[] {inputDir.toString(), outputDir.toString()});
+    createParentDirectories(inputFile);
+    writeFile(inputFile, readResource(String.format("%s/%s", TESTCONFIGS_PATH, inputFilename)));
+    main(new String[] {inputDir.toString(), outputDir.toString()});
 
     assertThat(
-        CommonUtil.readFile(outputFile),
-        equalTo(CommonUtil.readResource(String.format("%s/%s", TESTCONFIGS_PATH, outputFilename))));
+        readFile(outputFile),
+        equalTo(readResource(String.format("%s/%s", TESTCONFIGS_PATH, outputFilename))));
   }
 
   @Test
-  public void testMainValid() {
+  public void testMainValid() throws IOException {
     assertInputOutputPair("hierarchical", "flat");
   }
 
   @Test
-  public void testReplaceOnly() {
+  public void testReplaceOnly() throws IOException {
     assertInputOutputPair("replace_only", "replace_only_flattened");
   }
 
   @Test
-  public void testReplaceFilter() {
+  public void testReplaceFilter() throws IOException {
     assertInputOutputPair("replace_filter", "replace_filter_flattened");
   }
 }
